+++ srcs/expand_dollar.c
@@ -6,7 +6,7 @@
 /*   By: yohatana <yohatana@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/19 20:33:45 by yohatana          #+#    #+#             */
-/*   Updated: 2025/03/28 19:06:01 by yohatana         ###   ########.fr       */
+/*   Updated: 2025/03/30 13:21:04 by yohatana         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,8 @@
 
 static bool	expand_exec(t_token *curr, int *index, t_env *env);
 static void	skip_single_quote(t_token *curr, int *index);
-bool	replace_env_word(t_token *curr, int *index, t_env *env);
-char	*split_env_key(t_token *curr, int *index);
-bool	replace_word(t_token *curr, int *index, t_env replace);
-char	*split_after_word(t_token *curr, int *index, t_env replace);
+static bool	replace_env_word(t_token *curr, int *index, t_env *env);
+static char	*split_env_key(t_token *curr, int *index);
 
 bool	expand_dollar(t_token **head, t_env *env)
 {
@@ -40,7 +38,6 @@ bool	expand_dollar(t_token **head, t_env *env)
 		}
 		if (err_flg)
 			break ;
-		printf("word :%s\n", curr->word);
 		curr = curr->next;
 	}
 	if (err_flg)
@@ -70,7 +67,7 @@ static bool	expand_exec(t_token *curr, int *index, t_env *env)
 		}
 	}
 	else
-	err_flg = replace_env_word(curr, index, env);
+		err_flg = replace_env_word(curr, index, env);
 	return (err_flg);
 }
 
@@ -88,7 +85,7 @@ static void	skip_single_quote(t_token *curr, int *index)
 	*index = i;
 }
 
-bool	replace_env_word(t_token *curr, int *index, t_env *env)
+static bool	replace_env_word(t_token *curr, int *index, t_env *env)
 {
 	t_env	replace;
 
@@ -100,7 +97,7 @@ bool	replace_env_word(t_token *curr, int *index, t_env *env)
 		*index = *index + 1;
 		return (false);
 	}
-	replace.value = ft_strdup(search_env(replace.key, env)->value);
+	replace.value = ft_strdup(serch_env_value(replace.key, env));
 	if (replace_word(curr, index, replace))
 	{
 		free(replace.key);
@@ -110,7 +107,17 @@ bool	replace_env_word(t_token *curr, int *index, t_env *env)
 	return (false);
 }
 
-char	*split_env_key(t_token *curr, int *index)
+// char	*serch_env_value(char *key, t_env *env)
+// {
+// 	t_env	*find;
+
+// 	find = search_env(key, env);
+// 	if (find == NULL)
+// 		return ("");
+// 	return (find->value);
+// }
+
+static char	*split_env_key(t_token *curr, int *index)
 {
 	char	*key;
 	int		i;
@@ -125,38 +132,38 @@ char	*split_env_key(t_token *curr, int *index)
 	return (key);
 }
 
-bool	replace_word(t_token *curr, int *index, t_env replace)
-{
-	char	*after_token_word;
-	char	*temp;
-	char	*dollar_after_word;
+// bool	replace_word(t_token *curr, int *index, t_env replace)
+// {
+// 	char	*after_token_word;
+// 	char	*temp;
+// 	char	*dollar_after_word;
 
-	after_token_word = ft_calloc(ft_strlen(curr->word) - \
-						ft_strlen(replace.key) + ft_strlen(replace.value), \
-						sizeof(char));
-	if (!after_token_word)
-		return (true);
-	after_token_word = ft_substr(curr->word, 0, *index);
-	temp = ft_strjoin(after_token_word, replace.value);
-	free(after_token_word);
-	dollar_after_word = split_after_word(curr, index, replace);
-	if (!dollar_after_word)
-		return (true);
-	after_token_word = ft_strjoin(temp, dollar_after_word);
-	free(curr->word);
-	free(temp);
-	if (!after_token_word)
-		return (true);
-	curr->word = after_token_word;
-	return (false);
-}
+// 	after_token_word = ft_calloc(ft_strlen(curr->word) - \
+// 						ft_strlen(replace.key) + ft_strlen(replace.value), \
+// 						sizeof(char));
+// 	if (!after_token_word)
+// 		return (true);
+// 	after_token_word = ft_substr(curr->word, 0, *index);
+// 	temp = ft_strjoin(after_token_word, replace.value);
+// 	free(after_token_word);
+// 	dollar_after_word = split_after_word(curr, index, replace);
+// 	if (!dollar_after_word)
+// 		return (true);
+// 	after_token_word = ft_strjoin(temp, dollar_after_word);
+// 	free(curr->word);
+// 	free(temp);
+// 	if (!after_token_word)
+// 		return (true);
+// 	curr->word = after_token_word;
+// 	return (false);
+// }
 
-char	*split_after_word(t_token *curr, int *index, t_env replace)
-{
-	char	*split_after_word;
+// char	*split_after_word(t_token *curr, int *index, t_env replace)
+// {
+// 	char	*split_after_word;
 
-	split_after_word = ft_substr(curr->word, \
-								*index + ft_strlen(replace.key), \
-								ft_strlen(curr->word));
-	return (split_after_word);
-}
+// 	split_after_word = ft_substr(curr->word, \
+// 								*index + ft_strlen(replace.key), \
+// 								ft_strlen(curr->word));
+// 	return (split_after_word);
+// }
